--- app/src/main/java/com/example/project_flow_android/data/remote/CertificationApi.kt
@@ -1,20 +1,0 @@
-package com.example.project_flow_android.data.remote
-
-import com.example.project_flow_android.feature.CertificationRequest
-import com.example.project_flow_android.feature.PostCertificationRequest
-import io.reactivex.rxjava3.core.Single
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface CertificationApi {
-
-    @POST("/email/")
-    fun postCertification(@Body request: CertificationRequest): Single<Response<Void>>
-
-    @POST("/email/verifyCode")
-    fun checkCertification(@Body request: PostCertificationRequest): Single<Response<Void>>
-
-
-
-}
