--- app/src/main/java/com/example/project_flow_android/data/remote/LoginApi.kt
@@ -1,15 +1,0 @@
-package com.example.project_flow_android.data.remote
-
-import com.example.project_flow_android.feature.LoginRequest
-import com.example.project_flow_android.feature.LoginResponse
-import io.reactivex.rxjava3.core.Single
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.Header
-import retrofit2.http.POST
-
-interface LoginApi {
-
-    @POST("/auth/login")
-    fun login(@Body request: LoginRequest): Single<Response<LoginResponse>>
-}
