+++ app/src/main/java/com/example/project_flow_android/data/remote/LoginApiImpl.kt
@@ -2,6 +2,7 @@ package com.example.project_flow_android.data.remote
 
 import com.example.project_flow_android.feature.*
 import com.example.project_flow_android.network.ApiProvider
+import com.example.project_flow_android.network.ProjectFlowAPI
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.annotations.NonNull
 import io.reactivex.rxjava3.core.Single
@@ -10,7 +11,7 @@ import retrofit2.Response
 
 class LoginApiImpl {
 
-    private fun providerSignApi(): LoginApi = ApiProvider.RetroFitBuilder.create(LoginApi::class.java)
+    private fun providerSignApi(): ProjectFlowAPI = ApiProvider.RetroFitBuilder.create(ProjectFlowAPI::class.java)
 
     fun loginApi(request: LoginRequest):@NonNull Single<Response<LoginResponse>> = providerSignApi().login(request)
         .observeOn(AndroidSchedulers.mainThread())
