+++ app/src/main/java/com/example/project_flow_android/viewmodel/register/RegisterViewModel.kt
@@ -5,13 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.project_flow_android.data.SharedPreferenceStorage
 import com.example.project_flow_android.data.remote.SignApiImpl
-import com.example.project_flow_android.feature.CertificationRequest
 import com.example.project_flow_android.feature.RegisterRequest
-import retrofit2.Callback
-import com.example.project_flow_android.network.ApiProvider
-import com.example.project_flow_android.network.ProjectFlowAPI
-import retrofit2.Call
-import retrofit2.Response
 
 class RegisterViewModel(
     private val signApiImpl: SignApiImpl,
@@ -68,22 +62,26 @@ class RegisterViewModel(
         val userName = sharedPrefenceStorage.getInfo("userName")
         val userPhonenumber = sharedPrefenceStorage.getInfo("userPhone")
 
-        if (userPassword.value?.length.toString() >= 8.toString() || userPassword.value?.length.toString() < 20.toString()) {
+        if (userPassword.value?.length ?: 0 in 8..16) {
             if (userPassword.value!! == userRePassword.value!!) {
                 signApiImpl.registerApi(RegisterRequest(sharedPrefenceStorage.getInfo(userName),
                     sharedPrefenceStorage.getInfo(userEmail),
                     sharedPrefenceStorage.getInfo(userPhonenumber), userPassword.value!!))
-                    .subscribe { subscribe ->
-                        if (subscribe.isSuccessful) {
+                    .subscribe({
+                        if (it.isSuccessful) {
                             _finishRegister.value = true
                         } else {
-                            _changeComment_2.value = "회원가입에 실패하였습니다"
+                            _changeComment_2.value = it.body().toString()
                         }
-                    }
-            } else
+                    }, {
+                        _changeComment_2.value = "회원가입에 실패하였습니다"
+                    })
+            } else {
                 _changeComment_2.value = "작성 비밀번호가 일치하지 않습니다"
-        }
-        else
+            }
+        } else {
             _changeComment_2.value = "비밀번호를 8자에서 16자로 설정해주세요"
+        }
+
     }
 }
