+++ app/src/main/res/layout/activity_edit_password.xml
@@ -151,7 +151,7 @@
         android:id="@+id/comment2_tv"
         android:layout_width="wrap_content"
         android:layout_height="15dp"
-        android:layout_marginBottom="2dp"
+        android:layout_marginBottom="15dp"
         android:textColor="@color/red"
         android:text="@{vm.changeComment2}"
         app:layout_constraintBottom_toTopOf="@+id/register_btn"
